add original true space coordinate tracking

+++ starsystem.cpp
@@ -7,7 +7,7 @@ StarSystem::StarSystem():
 {
 }
 
-StarSystem::StarSystem(std::string_view name, float true_space_position_x, float true_space_position_y):
+StarSystem::StarSystem(std::string_view name, int true_space_position_x, int true_space_position_y):
     name_property{name},
     true_space_position_x_property{true_space_position_x},
     true_space_position_y_property{true_space_position_y}
@@ -19,12 +19,12 @@ std::string_view StarSystem::name() const
     return name_property;
 }
 
-float StarSystem::true_space_position_x() const
+int StarSystem::true_space_position_x() const
 {
     return true_space_position_x_property;
 }
 
-float StarSystem::true_space_position_y() const
+int StarSystem::true_space_position_y() const
 {
     return true_space_position_y_property;
 }
