+++ starsystem.h
@@ -7,15 +7,15 @@ class StarSystem
 {
 public:
     StarSystem();
-    StarSystem(std::string_view name, float true_space_position_x, float true_space_position_y);
+    StarSystem(std::string_view name, int true_space_position_x, int true_space_position_y);
     std::string_view name() const;
-    float true_space_position_x() const;
-    float true_space_position_y() const;
+    int true_space_position_x() const;
+    int true_space_position_y() const;
 
 private:
     std::string_view name_property;
-    float true_space_position_x_property;
-    float true_space_position_y_property;
+    int true_space_position_x_property;
+    int true_space_position_y_property;
 };
 
 #endif // STARSYSTEM_H
